add new info popup - password success
