+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import Header from './components/Header/Header';
-import PasswordSaved from './pages/PasswordSaved/PasswordSaved';
 import PopupFeedbackThanks from './pages/PopupFeedbackThanks/PopupFeedbackThanks';
 import NewPassword from './pages/NewPassword/NewPassword';
 import PasswordRecovery from './pages/PasswordRecovery/PasswordRecovery';
@@ -26,7 +25,6 @@ function App() {
                 <Route path="/signup" element={<SignUp />} />
                 <Route path="/recovery_pass" element={<PasswordRecovery />} />
                 <Route path="/new_pass" element={<NewPassword />} />
-                <Route path="/password_done" element={<PasswordSaved />} />
                 <Route path="/profile" element={<ProtectedRoute component={<Profile />} islogin={isLogin} />} />
                 <Route path="/feedback_done" element={<ProtectedRoute component={<PopupFeedbackThanks />} islogin={isLogin} />} />
             </Routes>
