+++ src/pages/NewPassword/PopupPasswordSaved/PopupPasswordSaved.module.scss
@@ -8,7 +8,7 @@
     &__success {
         width: 38px;
         height: 38px;
-        background-image: url(../../../../vendor/images/successful.png);
+        background-image: url(../../../vendor/images/successful.png);
         background-size: cover;
         background-repeat: no-repeat;
     }
